Reorganized Models + Handling error if Server is not found
