+++ mobile/lib/models/service_results/result.dart
